+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/InvoiceEditPanel.java
@@ -48,9 +48,8 @@ public final class InvoiceEditPanel extends EntityEditPanel {
   private final EntityPanel invoiceLinePanel;
 
   public InvoiceEditPanel(SwingEntityEditModel editModel, EntityPanel invoiceLinePanel) {
-    super(editModel);
+    super(editModel, config -> config.clearAfterInsert(false));
     this.invoiceLinePanel = invoiceLinePanel;
-    clearAfterInsert().set(false);
   }
 
   @Override