+++ demos/manual/src/main/java/is/codion/framework/demos/manual/notes/NotesDemo.java
@@ -131,7 +131,7 @@ public final class NotesDemo {
 
     private NoteEditPanel(NoteEditModel editModel) {
       // CLEAR is the only standard control we require, for clearing the UI
-      super(editModel, EditControl.CLEAR);
+      super(editModel, config -> config.editControls(EditControl.CLEAR));
     }
 
     @Override
@@ -205,10 +205,10 @@ public final class NotesDemo {
     private NotePanel(NoteModel noteModel) {
       super(noteModel,
               new NoteEditPanel(noteModel.editModel()),
-              new NoteTablePanel(noteModel.tableModel()));
-      // No need to include the default control buttons since
-      // we added the CLEAR control button to the edit panel
-      configure().includeControls(false);
+              new NoteTablePanel(noteModel.tableModel()), config -> config
+                      // No need to include the default control buttons since
+                      // we added the CLEAR control button to the edit panel
+                      .includeControls(false));
     }
 
     @Override