+++ framework/db-core/src/main/java/is/codion/framework/db/DefaultCopy.java
@@ -34,7 +34,7 @@ import java.util.Map;
 
 import static java.util.Objects.requireNonNull;
 
-final class DefaultCopyEntities implements Copy {
+final class DefaultCopy implements Copy {
 
   private final EntityConnection source;
   private final EntityConnection destination;
@@ -43,7 +43,7 @@ final class DefaultCopyEntities implements Copy {
   private final int batchSize;
   private final boolean includePrimaryKeys;
 
-  DefaultCopyEntities(DefaultBuilder builder) {
+  DefaultCopy(DefaultBuilder builder) {
     this.source = builder.source;
     this.destination = builder.destination;
     this.entityTypes.addAll(builder.entityTypes);
@@ -65,7 +65,7 @@ final class DefaultCopyEntities implements Copy {
       if (!includePrimaryKeys) {
         entities.forEach(Entity::clearPrimaryKey);
       }
-      new DefaultInsertEntities.DefaultBuilder(destination, entities.iterator())
+      new DefaultInsert.DefaultBuilder(destination, entities.iterator())
               .batchSize(batchSize)
               .execute();
     }
@@ -110,7 +110,7 @@ final class DefaultCopyEntities implements Copy {
     @Override
     public Builder condition(Condition condition) {
       if (!entityTypes.contains(requireNonNull(condition).entityType())) {
-        throw new IllegalArgumentException("CopyEntities.Builder does not contain entityType: " + condition.entityType());
+        throw new IllegalArgumentException("Copy.Builder does not contain entityType: " + condition.entityType());
       }
       this.conditions.put(condition.entityType(), condition);
       return this;
@@ -123,7 +123,7 @@ final class DefaultCopyEntities implements Copy {
 
     @Override
     public Copy build() {
-      return new DefaultCopyEntities(this);
+      return new DefaultCopy(this);
     }
   }
 }